backend/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
psycopg2_binary==2.9.6
py_avataaars==1.1.2
PyJWT==2.8.0
Requests==2.31.0
Werkzeug==2.3.6
